+++ CHANGELOG.md
@@ -51,6 +51,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `use_ref_ranges` alias for `use_alt`
   - New *kicad* format to mimic KiCad's internal BoM.
   - `right_digits` option to control the decimals for position fields (#739)
+  - New `sort_style` named `field` and `sort_field` option to sort by one or
+    more arbitrary fields.
 - Draw Fancy Stackup:
   - `border_thickness`: Thickness of the borders of stackup drawing and stackup
     table (#747)